+++ test/test_regression.py
@@ -12,12 +12,15 @@ from bs4 import BeautifulSoup
         "src/diagrams.html",
         "src/diagram-settings.html",
         "src/diagram-refs.html",
+        "src/diagram-resolver-data.html",
         "src/diagram-override.html",
         "src/diagram-opt.html",
         "src/objects.html",
         "src/refs.html",
         "src/audodoc-lexer.html",
         "src/audodoc-bison.html",
+        "src/audodoc-rule.html",
+        "src/root-rule.html",
     ],
 )
 def test_regression(app, src, file_regression):
@@ -36,9 +39,16 @@ def test_regression(app, src, file_regression):
 
 @pytest.mark.sphinx("html", testroot="doc")
 @pytest.mark.test_params(shared_result="test_regression")
-def test_invalid_yaml(app):
+def test_warnings(app):
     app.build()
     warnings = app.warning.getvalue()
+    assert "syntax:rule can't be used outside of a diagram" in warnings
+    assert (
+        "trying to document rule input as part of grammar WrongGrammar, "
+        "but the rule is defined in grammar Bison" in warnings
+    )
+    assert "'syntax:diagram' directive doesn't support option :name:" in warnings
+    assert "diagram descriptions can't have custom resolver_data" in warnings
     assert "can't parse syntax diagram description" in warnings
     assert "invalid-yaml.rst" in warnings
     assert "line 9," in warnings