+++ CHANGELOG.md
@@ -2,9 +2,22 @@
 
 ## [unreleased]
 
-## [1.0.1-post2]
+## [1.1.0]
 
-- Fix links after move.
+- Added `syntax:autorule`.
+
+- Fixed bugs related to propagating options to nested objects.
+
+- Autodoc-related options can be used with the `diagram` directive,
+  they'll be picked up by nested `autorule`, `lexer-diagram`, and `parser-diagram`.
+
+- Clarified documentation around the `:import:` option. It only has effect
+  on grammar declarations, despite documentation mentioning that it can be used
+  on diagrams.
+
+- Fixed links after move.
+
+- Releases are no longer uploaded to test version of PyPi.
 
 ## [1.0.1-post1]
 
@@ -26,8 +39,8 @@
 
 - Initial release.
 
-[unreleased]: https://github.com/sphinx-contrib/syntax/compare/v1.0.1-post2...HEAD
-[1.0.1-post2]: https://github.com/sphinx-contrib/syntax/compare/v1.0.1-post1...v1.0.1-post1
+[unreleased]: https://github.com/sphinx-contrib/syntax/compare/v1.1.0...HEAD
+[1.1.0]: https://github.com/sphinx-contrib/syntax/compare/v1.0.1-post1...v1.0.1-post1
 [1.0.1-post1]: https://github.com/sphinx-contrib/syntax/compare/v1.0.1...v1.0.1-post1
 [1.0.1]: https://github.com/sphinx-contrib/syntax/compare/v1.0.0-post1...v1.0.1
 [1.0.0-post1]: https://github.com/sphinx-contrib/syntax/compare/v1.0.0...v1.0.0-post1